pyproject.toml
[tool.poetry]
name = "portal-docs"
version = "1.1.0"
description = ""
authors = ["p0lygun <solankivibhakar82@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "generators", from="."}
]


[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.0"
python-box = "^7.0.1"
selenium = "^4.11.2"
webdriver-manager = "^4.0.0"
grequests = "^0.7.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
isort = "^5.12.0"
black = "^23.9.1"


[tool.poetry.group.dev.dependencies]
aiohttp = "^3.8.5"

[tool.poetry.scripts]
gen-docs =        "generators.blocks_documentation.__main__:generate"
gen-docs-json =   "generators.blocks_json.__main__:generate"
gen-map-layouts = "generators.map_layouts.__main__:generate"
gen-hugo-menu =   "generators.hugo_menu.__main__:generate"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
